moar

+++ sched_tuner.sh
@@ -565,11 +565,14 @@ done
 $FULL_NAME -a" > /system/etc/sched_tuner_task
 	chmod 755 /system/etc/sched_tuner_task
 	chmod 755 /init.rc
-	cp /init.rc $external/init.rc.bak
+	if [[ ! -f $external/init.rc.bak ]]; then
+		cp /init.rc $external/init.rc.bak
+	fi
 	echo "
+
 service sched_tuner_task /system/etc/sched_tuner_task
-		user root
-		oneshot" >> /init.rc
+     user root
+     oneshot" >> /init.rc
 }
 main(){
 	while true; do
